+++ modules/apps/rate-limiting/types/msgs.go
@@ -13,6 +13,19 @@ import (
 	clienttypes "github.com/cosmos/ibc-go/v11/modules/core/02-client/types"
 )
 
+// channelIDRegex is looser than channeltypes.IsValidChannelID, which also caps the sequence at 20 digits and a uint64.
+var channelIDRegex = regexp.MustCompile(`^channel-\d+$`)
+
+// validateChannelOrClientID validates an IBC channel identifier of the form 'channel-{N}' or a client identifier.
+func validateChannelOrClientID(channelOrClientID string) error {
+	if !channelIDRegex.MatchString(channelOrClientID) && !clienttypes.IsValidClientID(channelOrClientID) {
+		return errorsmod.Wrapf(sdkerrors.ErrInvalidRequest,
+			"invalid channel or client-id (%s), must be of the format 'channel-{N}' or a valid client-id", channelOrClientID)
+	}
+
+	return nil
+}
+
 var (
 	_ sdk.Msg = &MsgAddRateLimit{}
 	_ sdk.Msg = &MsgUpdateRateLimit{}
@@ -43,35 +56,15 @@ func (msg *MsgAddRateLimit) ValidateBasic() error {
 		return errorsmod.Wrapf(sdkerrors.ErrInvalidRequest, "invalid denom (%s)", msg.Denom)
 	}
 
-	matched, err := regexp.MatchString(`^channel-\d+$`, msg.ChannelOrClientId)
-	if err != nil {
-		return errorsmod.Wrapf(sdkerrors.ErrInvalidRequest, "unable to verify channel-id (%s)", msg.ChannelOrClientId)
-	}
-	if !matched && !clienttypes.IsValidClientID(msg.ChannelOrClientId) {
-		return errorsmod.Wrapf(sdkerrors.ErrInvalidRequest,
-			"invalid channel or client-id (%s), must be of the format 'channel-{N}' or a valid client-id", msg.ChannelOrClientId)
+	if err := validateChannelOrClientID(msg.ChannelOrClientId); err != nil {
+		return err
 	}
 
-	if msg.MaxPercentSend.GT(sdkmath.NewInt(100)) || msg.MaxPercentSend.LT(sdkmath.ZeroInt()) {
-		return errorsmod.Wrapf(sdkerrors.ErrInvalidRequest,
-			"max-percent-send percent must be between 0 and 100 (inclusively), Provided: %v", msg.MaxPercentSend)
-	}
-
-	if msg.MaxPercentRecv.GT(sdkmath.NewInt(100)) || msg.MaxPercentRecv.LT(sdkmath.ZeroInt()) {
-		return errorsmod.Wrapf(sdkerrors.ErrInvalidRequest,
-			"max-percent-recv percent must be between 0 and 100 (inclusively), Provided: %v", msg.MaxPercentRecv)
-	}
-
-	if msg.MaxPercentRecv.IsZero() && msg.MaxPercentSend.IsZero() {
-		return errorsmod.Wrapf(sdkerrors.ErrInvalidRequest,
-			"either the max send or max receive threshold must be greater than 0")
-	}
-
-	if msg.DurationHours == 0 {
-		return errorsmod.Wrapf(sdkerrors.ErrInvalidRequest, "duration can not be zero")
-	}
-
-	return nil
+	return Quota{
+		MaxPercentSend: msg.MaxPercentSend,
+		MaxPercentRecv: msg.MaxPercentRecv,
+		DurationHours:  msg.DurationHours,
+	}.Validate()
 }
 
 // ----------------------------------------------
@@ -97,35 +90,15 @@ func (msg *MsgUpdateRateLimit) ValidateBasic() error {
 		return errorsmod.Wrapf(sdkerrors.ErrInvalidRequest, "invalid denom (%s)", msg.Denom)
 	}
 
-	matched, err := regexp.MatchString(`^channel-\d+$`, msg.ChannelOrClientId)
-	if err != nil {
-		return errorsmod.Wrapf(sdkerrors.ErrInvalidRequest, "unable to verify channel-id (%s)", msg.ChannelOrClientId)
-	}
-	if !matched && !clienttypes.IsValidClientID(msg.ChannelOrClientId) {
-		return errorsmod.Wrapf(sdkerrors.ErrInvalidRequest,
-			"invalid channel or client-id (%s), must be of the format 'channel-{N}' or a valid client-id", msg.ChannelOrClientId)
+	if err := validateChannelOrClientID(msg.ChannelOrClientId); err != nil {
+		return err
 	}
 
-	if msg.MaxPercentSend.GT(sdkmath.NewInt(100)) || msg.MaxPercentSend.LT(sdkmath.ZeroInt()) {
-		return errorsmod.Wrapf(sdkerrors.ErrInvalidRequest,
-			"max-percent-send percent must be between 0 and 100 (inclusively), Provided: %v", msg.MaxPercentSend)
-	}
-
-	if msg.MaxPercentRecv.GT(sdkmath.NewInt(100)) || msg.MaxPercentRecv.LT(sdkmath.ZeroInt()) {
-		return errorsmod.Wrapf(sdkerrors.ErrInvalidRequest,
-			"max-percent-recv percent must be between 0 and 100 (inclusively), Provided: %v", msg.MaxPercentRecv)
-	}
-
-	if msg.MaxPercentRecv.IsZero() && msg.MaxPercentSend.IsZero() {
-		return errorsmod.Wrapf(sdkerrors.ErrInvalidRequest,
-			"either the max send or max receive threshold must be greater than 0")
-	}
-
-	if msg.DurationHours == 0 {
-		return errorsmod.Wrapf(sdkerrors.ErrInvalidRequest, "duration can not be zero")
-	}
-
-	return nil
+	return Quota{
+		MaxPercentSend: msg.MaxPercentSend,
+		MaxPercentRecv: msg.MaxPercentRecv,
+		DurationHours:  msg.DurationHours,
+	}.Validate()
 }
 
 // ----------------------------------------------
@@ -148,16 +121,7 @@ func (msg *MsgRemoveRateLimit) ValidateBasic() error {
 		return errorsmod.Wrapf(sdkerrors.ErrInvalidRequest, "invalid denom (%s)", msg.Denom)
 	}
 
-	matched, err := regexp.MatchString(`^channel-\d+$`, msg.ChannelOrClientId)
-	if err != nil {
-		return errorsmod.Wrapf(sdkerrors.ErrInvalidRequest, "unable to verify channel-id (%s)", msg.ChannelOrClientId)
-	}
-	if !matched && !clienttypes.IsValidClientID(msg.ChannelOrClientId) {
-		return errorsmod.Wrapf(sdkerrors.ErrInvalidRequest,
-			"invalid channel or client-id (%s), must be of the format 'channel-{N}' or a valid client-id", msg.ChannelOrClientId)
-	}
-
-	return nil
+	return validateChannelOrClientID(msg.ChannelOrClientId)
 }
 
 // ----------------------------------------------
@@ -180,14 +144,5 @@ func (msg *MsgResetRateLimit) ValidateBasic() error {
 		return errorsmod.Wrapf(sdkerrors.ErrInvalidRequest, "invalid denom (%s)", msg.Denom)
 	}
 
-	matched, err := regexp.MatchString(`^channel-\d+$`, msg.ChannelOrClientId)
-	if err != nil {
-		return errorsmod.Wrapf(sdkerrors.ErrInvalidRequest, "unable to verify channel-id (%s)", msg.ChannelOrClientId)
-	}
-	if !matched && !clienttypes.IsValidClientID(msg.ChannelOrClientId) {
-		return errorsmod.Wrapf(sdkerrors.ErrInvalidRequest,
-			"invalid channel or client-id (%s), must be of the format 'channel-{N}' or a valid client-id", msg.ChannelOrClientId)
-	}
-
-	return nil
+	return validateChannelOrClientID(msg.ChannelOrClientId)
 }